# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "six-northnode.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/movie/": {
      "post": {
        "tags": [
          "Movie"
        ],
        "summary": "新增電影資訊",
        "description": "新增電影資訊",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "新增電影資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "捍衛任務5"
                },
                "$imgs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "level": {
                  "type": "number",
                  "example": 1
                },
                "desc": {
                  "type": "string",
                  "example": "一個神隱少女的故事"
                },
                "time": {
                  "type": "number",
                  "example": 180
                },
                "actors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "videos": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "releaseData": {
                  "type": "string",
                  "example": "2023-05-01"
                }
              },
              "required": [
                "name",
                "level",
                "desc",
                "time",
                "releaseData"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "捍衛任務5"
                    },
                    "imgs": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "level": {
                      "type": "number",
                      "example": 1
                    },
                    "desc": {
                      "type": "string",
                      "example": "一個神隱少女的故事"
                    },
                    "time": {
                      "type": "number",
                      "example": 240
                    },
                    "actors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "videos": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "status": {
                      "type": "number",
                      "example": 1
                    },
                    "releaseData": {
                      "type": "string",
                      "example": "2023-05-01T00:00:00.000Z"
                    },
                    "createTime": {
                      "type": "string",
                      "example": "2023-05-03T14:26:35.378Z"
                    },
                    "_id": {
                      "type": "string",
                      "example": "64526f1bd4947558c2b2b135"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Movie"
        ],
        "summary": "獲取電影列表",
        "description": "獲取電影列表",
        "parameters": [
          {
            "name": "isRelease\r",
            "in": "query"
          },
          {
            "name": "id",
            "in": "query"
          },
          {
            "name": "isRelease",
            "description": "是否上檔",
            "type": "boolean",
            "default": true,
            "in": "query"
          },
          {
            "name": "name",
            "description": "電影名稱模糊搜尋",
            "type": "string",
            "default": "鋼鐵人",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "鋼鐵人"
                      },
                      "imgs": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "level": {
                        "type": "number",
                        "example": 0
                      },
                      "desc": {
                        "type": "string",
                        "example": "string"
                      },
                      "time": {
                        "type": "number",
                        "example": 90
                      },
                      "actors": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "videos": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "release_data": {
                        "type": "string",
                        "example": "yymmdd-hms"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/movie/{id}": {
      "patch": {
        "tags": [
          "Movie"
        ],
        "summary": "修改電影資訊",
        "description": "修改電影資訊",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "電影id"
          },
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "編輯電影資訊用",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "星際大戰5"
                },
                "imgs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "level": {
                  "type": "number",
                  "example": 1
                },
                "desc": {
                  "type": "string",
                  "example": "一個神隱少女的故事"
                },
                "time": {
                  "type": "number",
                  "example": 180
                },
                "actors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "videos": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "releaseData": {
                  "type": "string",
                  "example": "2023-05-04"
                }
              },
              "required": [
                "name",
                "level",
                "desc",
                "time",
                "releaseData"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳更新成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "6453b1112133d9d29db76c35"
                        },
                        "name": {
                          "type": "string",
                          "example": "神快人kkk人xx"
                        },
                        "imgs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "level": {
                          "type": "number",
                          "example": 2
                        },
                        "desc": {
                          "type": "string",
                          "example": "一個沙贊的故事ddd"
                        },
                        "time": {
                          "type": "number",
                          "example": 240
                        },
                        "actors": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "videos": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "number",
                          "example": 1
                        },
                        "releaseData": {
                          "type": "string",
                          "example": "2323-01-01T00:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Movie"
        ],
        "summary": "刪除一則電影資訊",
        "description": "刪除一則電影資訊",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "電影id"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "644c5f99474605096360548c"
                    },
                    "name": {
                      "type": "string",
                      "example": "神隱少女"
                    },
                    "imgs": {
                      "type": "array",
                      "items": {}
                    },
                    "level": {
                      "type": "number",
                      "example": 1
                    },
                    "desc": {
                      "type": "string",
                      "example": "一個神隱少女的故事"
                    },
                    "time": {
                      "type": "number",
                      "example": 125
                    },
                    "actors": {
                      "type": "array",
                      "items": {}
                    },
                    "videos": {
                      "type": "array",
                      "items": {}
                    },
                    "releaseData": {
                      "type": "string",
                      "example": "2023-05-01T12:00:00.000Z"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "id不存在",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "找不到電影"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
}